Move fuzzer time tracking to another module.
This is part of an effort to make fuzz_task easier to understand.
